File swap and integration with SplitInfo over SplitName
Changed the name of the file and the class to decrease verbosity. Since the class is apart of the HollowKnight namespace, it is redundant to prefix the name with "HollowKnight", and takes up much more space. We also integrated with SplitInfo, using C# 7's pattern matching in order to implement SplitInfo in a switch statement, almost identically to the previously used switch statement matching against a SplitName value.